refactor(transform): isolate static local planning
